bouncer.h

+++ src/client.c
@@ -701,7 +701,7 @@ static bool handle_client_work(PgSocket *client, PktHdr *pkt)
 		slog_info(client, "DBEELINE - client link is in ready state %d", client->link->ready);
 		slog_info(client, "DBEELINE - client idle in tx %d", client->link->idle_tx);
 		slog_info(client, "DBEELINE - current last tx time is %d", client->last_tx_timestamp);
-		client->last_tx_timestamp = (int)time(NULL);
+		client->last_tx_timestamp = time(NULL);
 		slog_info(client, "DBEELINE - client last tx time is now set to %d", client->last_tx_timestamp);
 	}
 	if(client->link && client->link->idle_tx) {
